move scenario info code to load-breach
